Add files via upload
This is the first commit of the react app.  The progress so far in this app: launch the app with npm start, the app will call a WebAPI via fetch(), the response is interrogated for 2 pieces of information (the product title and description), display the product title and description in the html page.